requirements.txt
chromadb
sentence-transformers
google-generativeai
python-dotenv
pandas
